# project/dist/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-commerce System",
        "description": "Sistema básico de cadastro de usuários, produtos e vendas.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/purchases": {
            "post": {
                "summary": "Cadastra uma nova compra",
                "tags": ["COMPRAS"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Purchase"
                            },
                            "examples": {
                                "Register": {
                                    "summary": "Cadastra uma nova compra",
                                    "value": {
                                        "user_id": "38c9c1a2-7c82-4eed-9880-4609bfb0ed41",
                                        "product_id": "87ac3c4f-6c31-4ece-b627-1e0961c062eb",
                                        "quantity": 5
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido cadastrado com sucesso",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Create": {
                                        "value": {
                                            "id": "4720c32a-9eee-4b63-8450-e175a536f924",
                                            "client_name": "João",
                                            "delivery_date": "2022-10-05T12:09:04.374Z",
                                            "items_list_id": [
                                                "1",
                                                "2",
                                                "3"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "É retornado quando um item da lista de ID's não é contrado",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Error": {
                                        "summary": "Error",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "item list with id '38c9c1a2-7c82-4eed-9880-4609bfb0ed41z' not found",
                                            "error": "Not Found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Retorna todos os COMPRAS cadastrados",
                "tags": ["COMPRAS"],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Search All": {
                                        "value": [
                                            {
                                                "id": "4720c32a-9eee-4b63-8450-e175a536f924",
                                                "client_name": "João",
                                                "delivery_date": "2022-10-05T12:09:04.374Z",
                                                "shopping_list": [
                                                    {
                                                        "product": "COLHER DESCARTÁVEL MASTER PRAFESTA BRANCA C/ 50UN",
                                                        "quantity": 3,
                                                        "price": 5.99
                                                    },
                                                    {
                                                        "product": "GARFO DESCARTÁVEL MASTER PRAFESTA CRYSTAL C/ 50UN",
                                                        "quantity": 2,
                                                        "price": 7.49
                                                    },
                                                    {
                                                        "product": "FACA DESCARTÁVEL MASTER PRAFESTA CRYSTAL C/ 50UN",
                                                        "quantity": 5,
                                                        "price": 8.99
                                                    }
                                                ],
                                                "total_amount": 77.9
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/purchases/{id}": {
            "get": {
                "summary": "Retorna um pedido específico de acordo com o ID informado",
                "tags": ["COMPRAS"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Register": {
                                        "summary": "Retorna um pedido específico",
                                        "value": {
                                            "id": "4720c32a-9eee-4b63-8450-e175a536f924",
                                            "client_name": "João",
                                            "delivery_date": "2022-10-05T12:09:04.374Z",
                                            "items_list_id": [
                                                "1",
                                                "2",
                                                "3"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "É retornado quando um pedido não é encontrado",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Error": {
                                        "summary": "Error",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "order with id '4720c32a-9skfjq23' not found",
                                            "error": "Not Found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Altera um pedido de acordo com o ID informado",
                "tags": ["COMPRAS"],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Purchase"
                            },
                            "examples": {
                                "Register": {
                                    "summary": "Altera um pedido",
                                    "value": {
                                        "client_name": "João",
                                        "delivery_date": "2022-10-05T12:09:04.374Z",
                                        "items_list_id": [
                                            "1",
                                            "2",
                                            "3"
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Register": {
                                        "summary": "Mostra o pedido após alteração",
                                        "value": {
                                            "id": "4720c32a-9eee-4b63-8450-e175a536f924",
                                            "client_name": "João",
                                            "delivery_date": "2022-10-05T12:09:04.374Z",
                                            "items_list_id": [
                                                "1",
                                                "3"
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "É retornado quando um produto não é encontrado",
                        "content": {
                            "application/json": {
                                "examples": {
                                    "Error": {
                                        "summary": "Error",
                                        "value": {
                                            "statusCode": 404,
                                            "message": "product with id '101' not found",
                                            "error": "Not Found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Purchase": {
                "type": "object",
                "properties": {
                    "user_id": {
                        "type": "string",
                        "description": "ID do Usuário"
                    },
                    "product_id": {
                        "type": "string",
                        "description": "ID do Produto"
                    },
                    "quantity": {
                        "type": "int",
                        "description": "Quantidade do Produto"
                    },
                    "total_price": {
                        "type": "float",
                        "description": "Preço Total do Produto"
                    }
                }
            },
...
        }
    }
}